Refactor: Optimize project structure and add unit tests
This commit introduces several improvements to your project:

1.  **Project Structure Optimization:**
    *   I refactored the `src/game.ts` file into multiple, more focused modules under new directories:
        *   `src/components` (e.g., `balloon.ts`)
        *   `src/core` (e.g., `player.ts`, `questions.ts`, `audio.ts`)
        *   `src/utils` (e.g., `domElements.ts`, `helpers.ts`)
        *   `src/gameController.ts` now orchestrates the game logic.
    *   This modular structure enhances readability, maintainability, and scalability.

2.  **Unit Testing Framework (Jest):**
    *   I installed and configured Jest, `ts-jest`, `@types/jest`, and `jest-environment-jsdom`.
    *   The `testEnvironment` is set to `jsdom` to support DOM-related features if needed for future tests.
    *   The `package.json` test script now executes `jest`.

3.  **Initial Unit Tests:**
    *   I added unit tests for core utility functions:
        *   `src/core/questions.ts` (for `generateQuestion`)
        *   `src/utils/helpers.ts` (for `shuffleArray`)
    *   These tests verify the correctness of these functions and establish a foundation for future test development.
    *   A `src/__tests__` directory now holds all test files.

4.  **Build Process:**
    *   I ensured your project builds successfully with `npm run build` after the refactoring.
    *   I addressed TypeScript strict null check errors (TS18047) to improve type safety.

These changes aim to follow best practices for web development, making your codebase more robust and easier to work with.
