+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "dist/game.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "build": "./node_modules/.bin/tsc"
   },
   "repository": {
@@ -19,7 +19,11 @@
   },
   "homepage": "https://github.com/iceman1212/webgame#readme",
   "devDependencies": {
+    "@types/jest": "^29.5.14",
     "@types/node": "^24.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^30.0.0",
+    "ts-jest": "^29.3.4",
     "typescript": "^5.8.3"
   }
 }
